+++ packages/framework/src/hooks/useRegisterBindings.ts
@@ -1,6 +1,6 @@
 import type { RefCallback } from "react";
 import { useEffect } from "react";
-import { get, isEmpty, isString, merge } from "lodash-es";
+import { get, isEmpty, isString, keys, merge } from "lodash-es";
 import isEqual from "react-fast-compare";
 import type { InvokableMethods } from "../state";
 import { useWidgetId } from "./useWidgetId";
@@ -15,10 +15,13 @@ export interface RegisterBindingsResult<T> {
   rootRef: RefCallback<never>;
 }
 
-export const useRegisterBindings = <T extends Record<string, unknown>>(
+export const useRegisterBindings = <T extends { [key: string]: unknown }>(
   values: T,
   id?: string,
   methods?: InvokableMethods,
+  options?: {
+    forceState?: boolean;
+  },
 ): RegisterBindingsResult<T> => {
   const testId = get(values, ["testId"]);
 
@@ -52,7 +55,7 @@ export const useRegisterBindings = <T extends Record<string, unknown>>(
 
     if (
       isEqual(newValues, widgetState?.values) &&
-      isEqual(methods, widgetState?.invokable.methods)
+      isEqual(keys(methods), keys(widgetState?.invokable.methods))
     ) {
       return;
     }
@@ -71,11 +74,19 @@ export const useRegisterBindings = <T extends Record<string, unknown>>(
     id,
   ]);
 
+  useEffect(() => {
+    if (options?.forceState === true) {
+      setWidgetState({
+        values: newValues,
+        invokable: { id: resolvedWidgetId, methods },
+      });
+    }
+  }, [options?.forceState]);
+
   const updatedValues = widgetState?.values ?? newValues;
-  const htmlAttributes = get(updatedValues, "htmlAttributes") as Record<
-    string,
-    string
-  >;
+  const htmlAttributes = get(updatedValues, "htmlAttributes") as {
+    [key: string]: string;
+  };
 
   const { rootRef } = useHtmlPassThrough(
     htmlAttributes,