+++ packages/runtime/src/widgets/Charts/BarChart.tsx
@@ -54,7 +54,7 @@ export const BarChart: React.FC<ChartProps> = (props) => {
           datasets: config?.data.datasets ?? [],
         }}
         options={getMergedOptions(options, values?.title, config?.options)}
-        plugins={[ChartDataLabels]}
+        plugins={[ChartDataLabels, ...(config?.plugins || [])]}
       />
     </div>
   );