+++ packages/runtime/src/widgets/Charts/DoughnutChart.tsx
@@ -1,5 +1,6 @@
 import { Doughnut } from "react-chartjs-2";
 import type { ChartOptions } from "chart.js";
+import ChartDataLabels from "chartjs-plugin-datalabels";
 import { get } from "lodash-es";
 import { useState } from "react";
 import { useRegisterBindings } from "@ensembleui/react-framework";
@@ -15,7 +16,7 @@ export const DoughnutChart: React.FC<ChartProps> = (props) => {
   const { id, config } = props;
 
   const [title, setTitle] = useState(config?.title);
-  const [labels, setLabels] = useState<string[]>(config?.data?.labels || []);
+  const [labels, setLabels] = useState<string[]>(config?.data.labels || []);
 
   const { values } = useRegisterBindings({ labels, title }, id, {
     setLabels,
@@ -26,9 +27,10 @@ export const DoughnutChart: React.FC<ChartProps> = (props) => {
     <Doughnut
       data={{
         labels: values?.labels,
-        datasets: config?.data?.datasets as ChartDataSets[],
+        datasets: config?.data.datasets as ChartDataSets[],
       }}
       options={getMergedOptions(options, values?.title, config?.options)}
+      plugins={[ChartDataLabels as any, ...(config?.plugins || [])]}
       style={{
         ...(get(props, "styles") as object),
       }}