+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @ensembleui/react-starter
 
+## 0.3.27
+
+### Patch Changes
+
+- Updated dependencies [19d2369]
+- Updated dependencies [c55632d]
+- Updated dependencies [57b6697]
+  - @ensembleui/react-framework@0.1.28
+  - @ensembleui/react-runtime@0.1.38
+
+## 0.3.26
+
+### Patch Changes
+
+- Updated dependencies [35901fe]
+- Updated dependencies [bb779a7]
+  - @ensembleui/react-runtime@0.1.37
+
+## 0.3.25
+
+### Patch Changes
+
+- Updated dependencies [136d130]
+- Updated dependencies [17efe4f]
+  - @ensembleui/react-framework@0.1.27
+  - @ensembleui/react-runtime@0.1.36
+
+## 0.3.24
+
+### Patch Changes
+
+- Updated dependencies [0e65e56]
+- Updated dependencies [392d9d1]
+  - @ensembleui/react-framework@0.1.26
+  - @ensembleui/react-runtime@0.1.35
+
+## 0.3.23
+
+### Patch Changes
+
+- Updated dependencies [de7d06b]
+  - @ensembleui/react-framework@0.1.25
+  - @ensembleui/react-runtime@0.1.34
+
 ## 0.3.22
 
 ### Patch Changes