+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @ensembleui/react-runtime
 
+## 0.1.38
+
+### Patch Changes
+
+- c55632d: Handle nullish label in DataColumn
+- 57b6697: Fix stale data in ensemble.invokeAPI
+- Updated dependencies [19d2369]
+- Updated dependencies [57b6697]
+  - @ensembleui/react-framework@0.1.28
+
+## 0.1.37
+
+### Patch Changes
+
+- 35901fe: add ability to pass DataColumn's label as a widget too
+- bb779a7: Re-render screen when path changes
+
+## 0.1.36
+
+### Patch Changes
+
+- 136d130: Add click callback to Button
+- 17efe4f: added plugins support in doughnutchart and linechart
+- Updated dependencies [136d130]
+  - @ensembleui/react-framework@0.1.27
+
+## 0.1.35
+
+### Patch Changes
+
+- 0e65e56: add onChange action in Checkbox widget
+- Updated dependencies [0e65e56]
+- Updated dependencies [392d9d1]
+  - @ensembleui/react-framework@0.1.26
+
+## 0.1.34
+
+### Patch Changes
+
+- de7d06b: Fix various issues with action hooks
+- Updated dependencies [de7d06b]
+  - @ensembleui/react-framework@0.1.25
+
 ## 0.1.33
 
 ### Patch Changes